Refactor so imports aren't used unless needed * Protect import statements from happening unless they are needed * Speeds up how fast `help` text is generated, for example.
Fixes #23
